Clippy警告を色々修正

+++ crates/media/ffmpeg/src/decoder/av_errors.rs
@@ -1,7 +1,7 @@
 use ffmpeg_sys_next as sys;
 
 pub(crate) fn averror_eagain() -> i32 {
-    -(libc::EAGAIN as i32)
+    -libc::EAGAIN
 }
 
 pub(crate) fn averror_eof() -> i32 {
@@ -13,5 +13,5 @@ pub(crate) fn ignore_send_packet_result(result: i32) -> bool {
         || result == averror_eagain()
         || result == averror_eof()
         || result == sys::AVERROR_INVALIDDATA
-        || result == -(libc::EINVAL as i32)
+        || result == -libc::EINVAL
 }
